requirements.txt
asgiref==3.10.0
django==5.2.7
django-rest-passwordreset==1.5.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
dotenv==0.9.9
pillow==12.0.0
pyjwt==2.10.1
python-dotenv==1.2.1
simplejwt==2.0.1
sqlparse==0.5.3
typing==3.10.0.0
